Moved code to utilities file
